Refactor common (#540)
* Move tensorflow layer definitions to a new file

* Move Scheduler from A2C utils to common schedules file

* Add missing definitions for legacy Scheduler

* Move tensorflow-related utilities tf_utils

* Move total_episode_reward_logger to tf_util

* Move get_by_index to ACER codes (only used by ACER)

* Move EpisodeStats to ACER codes (only used by ACER)

* Finish refactoring a2c/utils.py

* Refactor ppo2.py (move shared code elsewhere)

* Refactor replay buffer (mode out from deepq to commons)

* Remove shared function from SAC and TD3 (get_vars)

* Remove unused code

* Move flatten_lists to common file

* Fix imports in tests

* Add missing import to ACER

* Fix ACER dtype error

* Rename replay_buffer -> buffers

* Remove unused import

* Fix import in a test

* Move orphan method to more social circles

* Move PPO1/TRPO seg_gen to commons

* Update changelog

* Move SAC/TD3 policy code under more suitable tf_layers

* Update to new traj_seg_gen

* Add list of what was moved where
