+++ stable_baselines/acer/acer_simple.py
@@ -4,14 +4,33 @@ import warnings
 import numpy as np
 import tensorflow as tf
 from gym.spaces import Discrete, Box
+from collections import deque
 
 from stable_baselines import logger
-from stable_baselines.a2c.utils import batch_to_seq, seq_to_batch, Scheduler, EpisodeStats, \
-    get_by_index, check_shape, avg_norm, gradient_add, q_explained_variance, total_episode_reward_logger
+from stable_baselines.common.schedules import Scheduler
+from stable_baselines.common.tf_util import batch_to_seq, seq_to_batch, \
+    check_shape, avg_norm, gradient_add, q_explained_variance, total_episode_reward_logger
 from stable_baselines.acer.buffer import Buffer
 from stable_baselines.common import ActorCriticRLModel, tf_util, SetVerbosity, TensorboardWriter
 from stable_baselines.common.runners import AbstractEnvRunner
 from stable_baselines.common.policies import ActorCriticPolicy, RecurrentActorCriticPolicy
+
+
+# For ACER
+def get_by_index(input_tensor, idx):
+    """
+    Return the input tensor, offset by a certain value
+
+    :param input_tensor: (TensorFlow Tensor) The input tensor
+    :param idx: (int) The index offset
+    :return: (TensorFlow Tensor) the offset tensor
+    """
+    assert len(input_tensor.get_shape()) == 2
+    assert len(idx.get_shape()) == 1
+    idx_flattened = tf.range(0, input_tensor.shape[0], dtype=tf.int64) * input_tensor.shape[1] + idx
+    offset_tensor = tf.gather(tf.reshape(input_tensor, [-1]),  # flatten input
+                              idx_flattened)  # use flattened indices
+    return offset_tensor
 
 
 def strip(var, n_envs, n_steps, flat=False):
@@ -58,6 +77,64 @@ def q_retrace(rewards, dones, q_i, values, rho_i, n_envs, n_steps, gamma):
     qrets = qrets[::-1]
     qret = seq_to_batch(qrets, flat=True)
     return qret
+
+
+class EpisodeStats:
+    def __init__(self, n_steps, n_envs):
+        """
+        Calculates the episode statistics
+
+        :param n_steps: (int) The number of steps to run for each environment
+        :param n_envs: (int) The number of environments
+        """
+        self.episode_rewards = []
+        for _ in range(n_envs):
+            self.episode_rewards.append([])
+        self.len_buffer = deque(maxlen=40)  # rolling buffer for episode lengths
+        self.rewbuffer = deque(maxlen=40)  # rolling buffer for episode rewards
+        self.n_steps = n_steps
+        self.n_envs = n_envs
+
+    def feed(self, rewards, masks):
+        """
+        Update the latest reward and mask
+
+        :param rewards: ([float]) The new rewards for the new step
+        :param masks: ([float]) The new masks for the new step
+        """
+        rewards = np.reshape(rewards, [self.n_envs, self.n_steps])
+        masks = np.reshape(masks, [self.n_envs, self.n_steps])
+        for i in range(0, self.n_envs):
+            for j in range(0, self.n_steps):
+                self.episode_rewards[i].append(rewards[i][j])
+                if masks[i][j]:
+                    reward_length = len(self.episode_rewards[i])
+                    reward_sum = sum(self.episode_rewards[i])
+                    self.len_buffer.append(reward_length)
+                    self.rewbuffer.append(reward_sum)
+                    self.episode_rewards[i] = []
+
+    def mean_length(self):
+        """
+        Returns the average length of each episode
+
+        :return: (float)
+        """
+        if self.len_buffer:
+            return np.mean(self.len_buffer)
+        else:
+            return 0  # on the first params dump, no episodes are finished
+
+    def mean_reward(self):
+        """
+        Returns the average reward of each episode
+
+        :return: (float)
+        """
+        if self.rewbuffer:
+            return np.mean(self.rewbuffer)
+        else:
+            return 0
 
 
 class ACER(ActorCriticRLModel):
