+++ stable_baselines/common/schedules.py
@@ -106,3 +106,149 @@ class LinearSchedule(Schedule):
     def value(self, step):
         fraction = min(float(step) / self.schedule_timesteps, 1.0)
         return self.initial_p + fraction * (self.final_p - self.initial_p)
+
+
+def get_schedule_fn(value_schedule):
+    """
+    Transform (if needed) learning rate and clip range
+    to callable.
+
+    :param value_schedule: (callable or float)
+    :return: (function)
+    """
+    # If the passed schedule is a float
+    # create a constant function
+    if isinstance(value_schedule, (float, int)):
+        # Cast to float to avoid errors
+        value_schedule = constfn(float(value_schedule))
+    else:
+        assert callable(value_schedule)
+    return value_schedule
+
+
+def constfn(val):
+    """
+    Create a function that returns a constant
+    It is useful for learning rate schedule (to avoid code duplication)
+
+    :param val: (float)
+    :return: (function)
+    """
+
+    def func(_):
+        return val
+
+    return func
+
+
+# ================================================================
+# Legacy scheduler used by A2C, AKCTR and ACER
+# ================================================================
+
+def constant(_):
+    """
+    Returns a constant value for the Scheduler
+
+    :param _: ignored
+    :return: (float) 1
+    """
+    return 1.
+
+
+def linear_schedule(progress):
+    """
+    Returns a linear value for the Scheduler
+
+    :param progress: (float) Current progress status (in [0, 1])
+    :return: (float) 1 - progress
+    """
+    return 1 - progress
+
+
+def middle_drop(progress):
+    """
+    Returns a linear value with a drop near the middle to a constant value for the Scheduler
+
+    :param progress: (float) Current progress status (in [0, 1])
+    :return: (float) 1 - progress if (1 - progress) >= 0.75 else 0.075
+    """
+    eps = 0.75
+    if 1 - progress < eps:
+        return eps * 0.1
+    return 1 - progress
+
+
+def double_linear_con(progress):
+    """
+    Returns a linear value (x2) with a flattened tail for the Scheduler
+
+    :param progress: (float) Current progress status (in [0, 1])
+    :return: (float) 1 - progress*2 if (1 - progress*2) >= 0.125 else 0.125
+    """
+    progress *= 2
+    eps = 0.125
+    if 1 - progress < eps:
+        return eps
+    return 1 - progress
+
+
+def double_middle_drop(progress):
+    """
+    Returns a linear value with two drops near the middle to a constant value for the Scheduler
+
+    :param progress: (float) Current progress status (in [0, 1])
+    :return: (float) if 0.75 <= 1 - p: 1 - p, if 0.25 <= 1 - p < 0.75: 0.75, if 1 - p < 0.25: 0.125
+    """
+    eps1 = 0.75
+    eps2 = 0.25
+    if 1 - progress < eps1:
+        if 1 - progress < eps2:
+            return eps2 * 0.5
+        return eps1 * 0.1
+    return 1 - progress
+
+
+SCHEDULES = {
+    'linear': linear_schedule,
+    'constant': constant,
+    'double_linear_con': double_linear_con,
+    'middle_drop': middle_drop,
+    'double_middle_drop': double_middle_drop
+}
+
+
+class Scheduler(object):
+    def __init__(self, initial_value, n_values, schedule):
+        """
+        Update a value every iteration, with a specific curve.
+
+        This is a legacy version of schedules, originally defined
+        in a2c/utils.py. Used by A2C, ACER and ACKTR algorithms.
+
+        :param initial_value: (float) initial value
+        :param n_values: (int) the total number of iterations
+        :param schedule: (function) the curve you wish to follow for your value
+        """
+        self.step = 0.
+        self.initial_value = initial_value
+        self.nvalues = n_values
+        self.schedule = SCHEDULES[schedule]
+
+    def value(self):
+        """
+        Update the Scheduler, and return the current value
+
+        :return: (float) the current value
+        """
+        current_value = self.initial_value * self.schedule(self.step / self.nvalues)
+        self.step += 1.
+        return current_value
+
+    def value_steps(self, steps):
+        """
+        Get a value for a given step
+
+        :param steps: (int) The current number of iterations
+        :return: (float) the value for the current number of iterations
+        """
+        return self.initial_value * self.schedule(steps / self.nvalues)
