+++ stable_baselines/ppo1/pposgd_simple.py
@@ -10,11 +10,14 @@ from stable_baselines.common import Dataset, explained_variance, fmt_row, zipsam
     TensorboardWriter
 from stable_baselines import logger
 import stable_baselines.common.tf_util as tf_util
+from stable_baselines.common.tf_util import total_episode_reward_logger
 from stable_baselines.common.policies import ActorCriticPolicy
 from stable_baselines.common.mpi_adam import MpiAdam
 from stable_baselines.common.mpi_moments import mpi_moments
-from stable_baselines.trpo_mpi.utils import traj_segment_generator, add_vtarg_and_adv, flatten_lists
-from stable_baselines.a2c.utils import total_episode_reward_logger
+from stable_baselines.common.misc_util import flatten_lists
+from stable_baselines.common.runners import traj_segment_generator
+from stable_baselines.trpo_mpi.utils import add_vtarg_and_adv
+
 
 
 class PPO1(ActorCriticRLModel):
